compactor

+++ arguments.py
@@ -57,6 +57,6 @@ class TrainerArguments(Seq2SeqTrainingArguments):
     prefix_len: Optional[int] = field(default=10)
     overwrite_cache: Optional[bool] = field(default=False)
     trainable_params_percentage: Optional[float] = field(default=None)
-    reduction_factor: Optional[int] = field(default=32)
+    reduction_factor: Optional[float] = field(default=32)
     layer_name: Optional[str] = field(default=None)
     bias_name: Optional[str] = field(default=None)
